+++ static/app/components/events/interfaces/crashContent/exception/banners/stacktraceBanners.spec.tsx
